Modified snippets to use placeholders
VSCode would complain when launching the extension that the "New ScalaTest
suite"-snippet was very likely confusing snippet-variables and
snippet-placeholders. So I changed them over to use placeholders, as
that both silences the warning and makes the snippets more useful as you
can just tab through the bits that you need to type in yourself.